+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New #388: Add `beforeCheckbox()` and `afterCheckbox()` methods to `Checkbox` and `CheckboxList` fields (@vjik)
 - New #388: Add `beforeRadio()` and `afterRadio()` methods to `RadioList` field (@vjik)
+- Enh #386: Bump minimal `yiisoft/html` version to `^4.2` (@vjik)
 
 ## 1.5.2 March 20, 2026